Don't use a pointer to an interface
This is part of a cleanup for https://github.com/abcxyz/jvs/issues/116. In golang-jwt, Token is a struct. In jwx, it's an interface. Returning a pointer to an interface is generally frowned upon. This changes the interface to always be returned as an interface, but there needs to be a larger effort around reducing to a single jwt library.