Ensure logger exists on grpc context (#173)
This ensures that the logger that is configured during setup is propagated to the actual GRPC service. Without this, we get a "default" logger which doesn't have the right format or log level.

This also updates to the latest abcxyz/pkg version and updates services to print out their configuration at debug level on boot.