Switch to using a single binary (#245)
This is a pretty significant rearchitecture that moves away from publishing one-binary-per-service to publishing a full binary that contains the CLI and server components. It is capable of running CLI commands, servers, or both. This dramatically reduces the number of binaries and container images that we need to manage, and it reduces the cognitive complexity for customers who have to choose between the various binaries to download.