Deprecate client-lib
This removes the client-lib package and embeds the functionality directly into the generated protos. Originally, there was a goal of releasing the client independently from the server, but that proved to be complex and unnecessary. We undid that change, but kept the client separate.

This moves the logic into the generated protos where they belong, but maintains compatibility by aliasing the types and functions inside client-lib.